refactor: move service function to services directory
